# src/main/java/com/probestack/forgestudio/design/api/AccountsApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.2.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.probestack.forgestudio.design.api;

import com.probestack.forgestudio.design.model.Account;
import com.probestack.forgestudio.design.model.CreateAccountRequest;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-05-08T17:04:28.442050077Z[GMT]")@Validated
@Tag(name = "Accounts", description = "the Accounts API")
public interface AccountsApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * POST /accounts : Create a new account
     *
     * @param createAccountRequest  (required)
     * @return Account created (status code 201)
     */
    @Operation(
        operationId = "createAccount",
        summary = "Create a new account",
        tags = { "Accounts" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Account created", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Account.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/accounts",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<Account> createAccount(
        @Parameter(name = "CreateAccountRequest", description = "", required = true) @Valid @RequestBody CreateAccountRequest createAccountRequest
    ) {
                        getRequest().ifPresent(request -> {
                    for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                        if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                            String exampleString = "{ \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\", \"balance\" : 15420.75, \"accountType\" : \"SAVINGS\", \"currency\" : \"USD\", \"id\" : \"ACC123456\", \"customerName\" : \"John Doe\" }";
                            ApiUtil.setExampleResponse(request, "application/json", exampleString);
                            break;
                        }
                    }
                });
                return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
                                                            }


    /**
     * GET /accounts/{accountId} : Get account details
     *
     * @param accountId  (required)
     * @return Account details (status code 200)
     */
    @Operation(
        operationId = "getAccount",
        summary = "Get account details",
        tags = { "Accounts" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Account details", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Account.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/accounts/{accountId}",
        produces = { "application/json" }
    )
    
    default ResponseEntity<Account> getAccount(
        @Parameter(name = "accountId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("accountId") String accountId
    ) {
                        getRequest().ifPresent(request -> {
                    for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                        if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                            String exampleString = "{ \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\", \"balance\" : 15420.75, \"accountType\" : \"SAVINGS\", \"currency\" : \"USD\", \"id\" : \"ACC123456\", \"customerName\" : \"John Doe\" }";
                            ApiUtil.setExampleResponse(request, "application/json", exampleString);
                            break;
                        }
                    }
                });
                return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
                                                            }


    /**
     * GET /accounts : List customer accounts
     *
     * @return Successful response (status code 200)
     */
    @Operation(
        operationId = "listAccounts",
        summary = "List customer accounts",
        tags = { "Accounts" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Successful response", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = Account.class)))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/accounts",
        produces = { "application/json" }
    )
    
    default ResponseEntity<List<Account>> listAccounts(
        
    ) {
                        getRequest().ifPresent(request -> {
                    for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                        if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                            String exampleString = "[ { \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\", \"balance\" : 15420.75, \"accountType\" : \"SAVINGS\", \"currency\" : \"USD\", \"id\" : \"ACC123456\", \"customerName\" : \"John Doe\" }, { \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\", \"balance\" : 15420.75, \"accountType\" : \"SAVINGS\", \"currency\" : \"USD\", \"id\" : \"ACC123456\", \"customerName\" : \"John Doe\" } ]";
                            ApiUtil.setExampleResponse(request, "application/json", exampleString);
                            break;
                        }
                    }
                });
                return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
                                                            }

}