# src/main/java/com/probestack/forgestudio/design/api/PaymentsApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.2.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.probestack.forgestudio.design.api;

import com.probestack.forgestudio.design.model.PaymentRequest;
import com.probestack.forgestudio.design.model.PaymentResponse;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-05-08T17:04:28.442050077Z[GMT]")@Validated
@Tag(name = "Payments", description = "the Payments API")
public interface PaymentsApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * POST /payments : Initiate payment
     *
     * @param paymentRequest  (required)
     * @return Payment processed (status code 201)
     */
    @Operation(
        operationId = "createPayment",
        summary = "Initiate payment",
        tags = { "Payments" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Payment processed", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = PaymentResponse.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/payments",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<PaymentResponse> createPayment(
        @Parameter(name = "PaymentRequest", description = "", required = true) @Valid @RequestBody PaymentRequest paymentRequest
    ) {
                        getRequest().ifPresent(request -> {
                    for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                        if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                            String exampleString = "{ \"paymentId\" : \"paymentId\", \"processedAt\" : \"2000-01-23T04:56:07.000+00:00\", \"status\" : \"PENDING\" }";
                            ApiUtil.setExampleResponse(request, "application/json", exampleString);
                            break;
                        }
                    }
                });
                return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
                                                            }

}