# src/main/java/com/probestack/forgestudio/design/api/ReportsApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.2.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.probestack.forgestudio.design.api;

import com.probestack.forgestudio.design.model.BalanceSheet;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-05-08T17:04:28.442050077Z[GMT]")@Validated
@Tag(name = "Reports", description = "the Reports API")
public interface ReportsApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * GET /reports/balance-sheet : Generate balance sheet report
     *
     * @param fiscalYear  (required)
     * @return Balance sheet report (status code 200)
     */
    @Operation(
        operationId = "getBalanceSheet",
        summary = "Generate balance sheet report",
        tags = { "Reports" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Balance sheet report", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = BalanceSheet.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/reports/balance-sheet",
        produces = { "application/json" }
    )
    
    default ResponseEntity<BalanceSheet> getBalanceSheet(
        @NotNull @Parameter(name = "fiscalYear", description = "", required = true, in = ParameterIn.QUERY) @Valid @RequestParam(value = "fiscalYear", required = true) Integer fiscalYear
    ) {
                        getRequest().ifPresent(request -> {
                    for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                        if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                            String exampleString = "{ \"assets\" : 6.027456183070403, \"liabilities\" : 1.4658129805029452, \"fiscalYear\" : 0, \"equity\" : 5.962133916683182 }";
                            ApiUtil.setExampleResponse(request, "application/json", exampleString);
                            break;
                        }
                    }
                });
                return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
                                                            }

}